refactor SearchPageTabbed filters into SearchFilters component
